requirements.txt
torch
matplotlib
ipykernel
torchvision
pillow 
tqdm
streamlit
